move jest event related utils to their own module
