+++ packages/jest-mock-environment/src/testEventUtils.js
@@ -1,0 +1,18 @@
+const isTestStartEvent = (event) => event.name === 'test_fn_start';
+const isTestsEndEvent = (event) => event.name === 'teardown';
+const getTestID = (event) => {
+  let id = event.name;
+
+  if (event.parent && event.parent.name !== 'ROOT_DESCRIBE_BLOCK') {
+    const parentID = getTestID(event.parent);
+    id = `${parentID}/${id}`;
+  }
+
+  return id;
+};
+
+export {
+  isTestStartEvent,
+  isTestsEndEvent,
+  getTestID
+};
